requirements.txt
# Dependências principais da aplicação
openai>=1.12.0
twilio>=8.10.0
requests>=2.31.0
boto3>=1.34.0
python-dotenv>=1.0.0

